Sacando warnings

+++ TPE/Userland/SampleCodeModule/snake.c
@@ -432,7 +432,7 @@ void showScore(){
 }
 
 int getseconds(){
-  uint64_t currTime[6];
+  int currTime[6];
   sys_time(currTime);
 
   // printf("%d%d , %d%d, %d%d",currTime[5],currTime[4],currTime[3],currTime[2],currTime[1],currTime[0]);
